+++ messages/cli.py
@@ -111,8 +111,8 @@ def main_email(ctx, **kwds):
 @main.command("twilio")
 @click.argument("profile", type=click.STRING, required=False)
 @click.argument("body", type=click.STRING, default="", required=False)
-@option("-f", "--from", "from_", help="Originating (from) phone number")
-@option("-t", "--to", multiple=True, help="Recipient phone numbers.")
+@option("-f", "--from", "from_", help="Originating (from) phone number.")
+@option("-t", "--to", multiple=True, help="Recipient phone number.")
 @option("-m", "--file", help="Read message body from filepath.")
 @option(
     "-a",
@@ -234,3 +234,36 @@ def main_telegram(ctx, **kwds):
     * Example: messages telegram myTelegramProfile "Hello from Telegram." -a "https://somefile.jpg" --verbose
     """
     send_message("telegrambot", kwds)
+
+
+@main.command("whatsapp")
+@click.argument("profile", type=click.STRING, required=False)
+@click.argument("body", type=click.STRING, default="", required=False)
+@option("-f", "--from", "from_", help="Originating (from) phone number.")
+@option("-t", "--to", multiple=True, help="Recipient phone number.")
+@option("-m", "--file", help="Read message body from filepath.")
+@option(
+    "-a",
+    "--attach",
+    "attachments",
+    multiple=True,
+    help="Attachments -- url for image to attach.",
+)
+@option("-S", "--save", is_flag=True, help="Save default values in current profile.")
+@option(
+    "-V",
+    "--verbose",
+    is_flag=True,
+    help="Display verbose output and debug information.",
+)
+@click.pass_context
+def main_whatsapp(ctx, **kwds):
+    """Send WhatsApp text message via the Twilio API.
+
+    * [PROFILE]: Pre-configured user profile.
+
+    * [BODY]:    Message body text.
+
+    * Example: messages whatsapp myWhatsAppProfile 'hello from whatsapp' -t '+12223334444' --verbose
+    """
+    send_message("whatsapp", kwds)